Enable use of Raysect in interactive Jupyter notebooks
When using the `%matplotlib notebook` magic in Jupyter (iPython)
notebooks, the `pause()` function is not implemented. This causes
Matplotlib to raise an error when observations are made using
certain pipelines (RGB, bayer, spectral and mono pipelines).

However, the `pause()` function is only called to work around a lack
of live updating in the Matplotlib QT backend. This workaround is
not needed in the notebook backend, so can be skipped.