Add functions for max, min and peak-to-peak values of memoryviews
The Python built-in min() and max() functions are very slow on
memoryviews, and Numpy's min and max are also slow when called many
times on small memoryviews (e.g. coordinate arrays). This commit adds
functions which operate at the C level to quickly calculate the max,
min and peak-to-peak values of a memoryview to a 1D array of type
double.