Propagate exceptions in worker processes to caller
Ensure that exceptions in worker processes do not cause the entire
program to hang. The consumer process in the Multicore engine expects
every task to put something in the results queue, so we either put the
result itself (if the task ran without error) or the exception raised
by the task. This means the consumer is never stuck waiting for
results in the queue which will never arrive. A calling program can
catch the resulting RuntimeError if any of the workers fails, and deal
with it accordingly.