Add vector functions to function framework.
Initial support is for VectorFunction2D. It's based on Function2D, but
when evaluated these functions return a Vector3D instead of a
double. Constant and auto wrapping are also provided. Arithmetic
methods are provided, but only for operands which are supported by
Vector3D cython methods (add, sub, mul, div).