Build in parallel by default
When installing with pip or PyPA's `build` utility, , the `-j` option
for parallel builds is not passed by default. This increases the build
time on systems where wheels need to be built, e.g. for releases or
on systems where no pre-built wheel is available.

This commit introduces a modified `build_ext` command which by default
parallelises over all available CPUs, with an optional override using
the `RAYSECT_BUILD_JOBS` environment variable if for any reason the
number of build processes needs to be controlled manually.