Update notes and scripts for building wheels
Wheels for Python 3.7 to 3.9 can be built using manylinux2010, as
there are binary wheels of oldest-supported-numpy available for these
versions. For 3.10 onwards manylinux2014 is required for numpy binary
wheels. Building numpy from source in the manylinux containers is not
viable.

There are now two scripts, one for each manylinux version, which will
build wheels for all upstream-supported Python versions. These can be
run as-is: info is in the first few lines of these files. Also, the
dev notes on building wheels wtih manylinux have been updated to
reference these files.