Implement hybrid render engine
This render engine uses a combination of distributed and shared memory
tools to perform parallel rendering. MPI (mesage passing interface) is
used to communicate between processes with separate memory, for
example on different nodes of a compute cluster. Within each MPI
process a secondary render engine is used for shared memory
computation.

The secondary (or sub) render engine may be any other render
engine. `SerialEngine` and `MulticoreEngine` are currently supported,
though the former is less efficient than using the MPIEngine due to
additional IPC overhead. The engine+subengine architecture leaves room
to expand to additional engines in future, such as a
`MultiThreadEngine` in a free-threaded (no GIL) Python build.

Gathering of render results is done in the main process on MPI rank
0. A sub process is spawned on rank 0 in which the sub render engine
runs. For this to work efficiently the operating system must support
cheap process forking without memcopy. At present only Linux is known
to support this properly. MacOS may work if you're lucky. This
limitation means the render engine is not truly cross platform: if
cross-platform compatibility is required then SerialEngine and
MPIEngine are still the only suitable choices.

Ideally the result gathering and updates should not require spawning a
new process: a separate thread would make more logical sense. But
doing it in the same thread as the sub-engine is run causes resource
contention and therefore significant performance degradation.
Suggestions for fixing this would be most helpful.

Hybrid MPI+multiprocessing has many gotchas, and care must be taken
when launching the program especially with a job scheduler. The
documentation includes some hints about how the program should be run
with a few popular schedulers, and there is also a static helper
method for getting an appropriate number of processes for a parallel
sub worker. But ultimately it's the responsibility of the end user to
configure the engine and start the job properly: the render engine
itself does very little of this automatically.

2 new demos have been added showcasing this new render engine: one to
render the Raysect logo and one doing a high fidelity, multi-pass
adaptive sampling render of the Cornell Box.